+++ clock/SyncClockManager.h
@@ -31,11 +31,11 @@ public:
     ~SyncClockManager();
 
     SyncState get_current_video_sync_state(
-        double next_pts, AVRational & time_base, int serial, double * remaining_time, 
+        double next_pts, AVRational & time_base, int serial, long long * remaining_time, 
         double current_time = av_gettime_relative() / 1000000.0);
 
     SyncState get_current_audio_sync_state(
-        double next_pts, AVRational & time_base, int serial, double * remaining_time,
+        double next_pts, AVRational & time_base, int serial, long long * remaining_time,
         double current_time = av_gettime_relative() / 1000000.0);
 
     void seek(int64_t position);
